ai_assignment1_2026/requirements.txt
commonroad-io==2021.1
tqdm==4.50.2
networkx==2.4
ipywidgets==7.5.1
ipython-autotime==0.1
matplotlib==3.3.4        
numpy==1.19.5     
ipython==7.18.1
pyyaml==5.3.1
imageio==2.9.0
shapely==1.7.1  
setuptools==42.0.1